requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.11.1
certifi==2021.5.30
cffi==1.14.6
cloudinary==1.26.0
constantly==15.1.0
cryptography==3.4.8
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.8.0
django-phone-field==1.8.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
Pillow==9.1.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
python-http-client==3.3.7
pytz==2021.1
sendgrid==6.9.7
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
starkbank-ecdsa==2.0.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
zope.interface==5.4.0
